rename plugin, cleanup code, enable user API
