+++ docs/reference/api.rst
@@ -132,7 +132,7 @@ Plugins
 .. automodapi:: jdaviz.configs.imviz.plugins.links_control.links_control
    :no-inheritance-diagram:
 
-.. automodapi:: jdaviz.configs.imviz.plugins.rotate_image.rotate_image
+.. automodapi:: jdaviz.configs.imviz.plugins.rotate_canvas.rotate_canvas
    :no-inheritance-diagram:
 
 .. automodapi:: jdaviz.configs.mosviz.plugins.row_lock.row_lock
