+++ jdaviz/configs/imviz/imviz.yaml
@@ -27,7 +27,7 @@ tray:
   - imviz-line-profile-xy
   - imviz-aper-phot-simple
   - imviz-catalogs
-  - imviz-rotate-image
+  - imviz-rotate-canvas
   - g-export-plot
 viewer_area:
   - container: col
