+++ jdaviz/configs/imviz/plugins/rotate_canvas/__init__.py
@@ -1,0 +1,1 @@
+from .rotate_canvas import *  # noqa
