--- jdaviz/configs/imviz/plugins/rotate_image/__init__.py
@@ -1,1 +1,0 @@
-from .rotate_image import *  # noqa
